criando modulo e usando o script
