+++ index.html
@@ -8,10 +8,10 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="stylesheet" href="css/table.css">
   <link rel="stylesheet" href="css/table-dark.css">
+  <script type="module" src="js/scripts.js" defer></script>
+  <script nomodule src="js/data.js" defer></script>
   <script src="js/scripts-layout.js" defer></script>
-  <script src="js/scripts.js" defer></script>
-  <script src="js/data.js" defer></script>
-  </head>
+</head>
 <body class="">
   <header>
     <a href="#"><h1>Criar novo CCO-Informa</h1></a>
